Renamed some more files and simplified the spline utils
